requirements.txt
azure-ai-ml
azure-identity
mltable